requirements.txt
appdirs==1.4.0
configparser==3.5.0
Django==1.10.5
django-contrib-comments==1.8.0
django-photologue==3.6
django-sortedm2m==1.3.3
ExifRead==2.1.2
mysqlclient==1.3.9
olefile==0.44
packaging==16.8
Pillow==4.0.0
pycrypto==2.6.1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
vobject==0.9.4.1
